+++ Game haha/Assets/Scripts/scr_DamageNumber.cs
@@ -24,7 +24,7 @@ public class scr_DamageNumber : MonoBehaviour
         if (durationtimer >= duration) { Destroy(gameObject); } else { durationtimer += Time.fixedDeltaTime; }
     }
 
-    public void SetNumber(float number, int order)
+    public void SetNumber(int number, int order)
     {
         GetComponent<TextMeshPro>().SetText(number.ToString("n0"));
         GetComponent<TextMeshPro>().sortingOrder = order;
